+++ frontend/src/components/Calendar/Calendar.tsx
@@ -1,39 +1,21 @@
-import { useState } from 'react';
 import { css } from 'styled-components';
-import { arrayOf } from '~/utils/arrayOf';
-import { parseDate } from '~/utils/parseDate';
 import Text from '~/components/common/Text/Text';
 import Button from '~/components/common/Button/Button';
 import DateCell from '~/components/Calendar/DateCell/DateCell';
 import { ArrowLeftIcon, ArrowRightIcon } from '~/assets/svg';
 import * as S from './Calendar.styled';
+import useCalendar from '~/hooks/useCalendar';
 
 const DAYS_OF_WEEK = ['일', '월', '화', '수', '목', '금', '토'] as const;
 
 const Calendar = () => {
-  const [currentDate, setCurrentDate] = useState(new Date());
-  const { year, month } = parseDate(currentDate);
-  const { day: startDayOfMonth } = parseDate(new Date(currentDate.setDate(1)));
+  const {
+    year,
+    month,
+    calendar,
 
-  const handlePrevButtonClick = () => {
-    setCurrentDate(() => new Date(year, month - 1));
-  };
-
-  const handleNextButtonClick = () => {
-    setCurrentDate(() => new Date(year, month + 1));
-  };
-
-  const createCalendar = () =>
-    arrayOf(6).map((weekIndex) =>
-      arrayOf(7).map((dayIndex) => {
-        const day = weekIndex * 7 + dayIndex - startDayOfMonth + 1;
-        const date = new Date(year, month, day);
-
-        return date;
-      }),
-    );
-
-  const calendar = createCalendar();
+    handlers: { handlePrevButtonClick, handleNextButtonClick },
+  } = useCalendar();
 
   return (
     <S.Container>
